Move class only used in tests to tests
