+++ src/ert/ensemble_evaluator/snapshot.py
@@ -525,72 +525,6 @@ class SnapshotDict(BaseModel):
     metadata: Dict[str, Any] = {}
 
 
-class SnapshotBuilder(BaseModel):
-    steps: Dict[str, Step] = {}
-    metadata: Dict[str, Any] = {}
-
-    def build(
-        self,
-        real_ids: Sequence[str],
-        status: Optional[str],
-        start_time: Optional[datetime.datetime] = None,
-        end_time: Optional[datetime.datetime] = None,
-    ) -> Snapshot:
-        top = SnapshotDict(status=status, metadata=self.metadata)
-        for r_id in real_ids:
-            top.reals[r_id] = RealizationSnapshot(
-                active=True,
-                steps=self.steps,
-                start_time=start_time,
-                end_time=end_time,
-                status=status,
-            )
-        return Snapshot(top.dict())
-
-    def add_step(
-        self,
-        step_id: str,
-        status: Optional[str],
-        start_time: Optional[datetime.datetime] = None,
-        end_time: Optional[datetime.datetime] = None,
-    ) -> "SnapshotBuilder":
-        self.steps[step_id] = Step(
-            status=status, start_time=start_time, end_time=end_time
-        )
-        return self
-
-    def add_job(  # pylint: disable=too-many-arguments
-        self,
-        step_id: str,
-        # This is not tested in test_snapshot.py..
-        job_id: str,
-        index: str,
-        name: Optional[str],
-        status: Optional[str],
-        data: Optional[Dict[str, Any]],
-        start_time: Optional[datetime.datetime] = None,
-        end_time: Optional[datetime.datetime] = None,
-        stdout: Optional[str] = None,
-        stderr: Optional[str] = None,
-    ) -> "SnapshotBuilder":
-        step = self.steps[step_id]
-        step.jobs[job_id] = Job(
-            status=status,
-            index=index,
-            data=data,
-            start_time=start_time,
-            end_time=end_time,
-            name=name,
-            stdout=stdout,
-            stderr=stderr,
-        )
-        return self
-
-    def add_metadata(self, key: str, value: Any) -> "SnapshotBuilder":
-        self.metadata[key] = value
-        return self
-
-
 def _from_old_super_nested_dict(data: Mapping[str, Any]) -> PartialSnapshot:
     partial = PartialSnapshot()
     if "metadata" in data:
